+++ apps/examples/wc-next13/pages/index.tsx
@@ -1,22 +1,20 @@
 import React, { useState, useRef } from 'react';
 import Head from 'next/head';
-import { ButtonProps, sizes, variants } from '@justeattakeaway/pie-button';
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import { PieButton } from '@justeattakeaway/pie-button/dist/react';
+import {
+    PieButton, ButtonProps, sizes, variants,
+} from '@justeattakeaway/pie-button/dist/react';
 import '@justeattakeaway/pie-css';
 
 export default function Home () {
     const variantIndex = useRef(0);
     const [count, setCount] = useState(0);
-    const [variantName, setVariantName] = useState('primary');
+    const [variantName, setVariantName] = useState<ButtonProps['variant']>('primary');
 
     const increment = () => setCount(count + 1);
     const decrement = () => setCount(count - 1);
     const switchVariant = () => {
         variantIndex.current += 1;
-        const variantsKeys: readonly string[] = variants;
-        const variant: ButtonProps['variant'] = variants[variantIndex.current % variantsKeys.length];
+        const variant: ButtonProps['variant'] = variants[variantIndex.current % variants.length];
 
         setVariantName(variant);
     };