+++ packages/tools/generator-pie-component/src/app/templates/src/index.ts
@@ -4,13 +4,11 @@ import styles from './<%= fileName %>.scss?inline';
 import { <%= componentName %>Props } from './defs';
 
 // Valid values available to consumers
-export {
-    type <%= componentName %>Props,
-};
+export * from './defs';
 
 const componentSelector = 'pie-<%= fileName %>';
 
-<% if (needsRTL) { %>export class Pie<%= componentName %> extends RtlMixin(LitElement) implements <%= componentName %>Props {<% } 
+<% if (needsRTL) { %>export class Pie<%= componentName %> extends RtlMixin(LitElement) implements <%= componentName %>Props {<% }
 else { %>export class Pie<%= componentName %> extends LitElement implements <%= componentName %>Props {<% } %>
     render () {
         return html`<h1 data-test-id="pie-<%= fileName %>">Hello world!</h1>`;