+++ packages/components/pie-toggle-switch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @justeattakeaway/pie-toggle-switch
 
+## 0.4.0
+
+### Minor Changes
+
+- [Updated] - Export all of `defs.ts` file from component `index.ts` files ([#722](https://github.com/justeattakeaway/pie/pull/722)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
 ## 0.3.0
 
 ### Minor Changes